pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flipformat"
version = "1.1.0"
description = "Create, read, and view .flip dual-sided image files."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "pytesseract>=0.3.10",
    "pillow-heif>=0.16.0",
]

[project.scripts]
flip = "flipformat.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
